Found I could not build both debug and timestamp modes. Fixed CFLAGS, including adding automake default CC flags. This exposed many warnings which had to get cleaned up.

+++ src/tcpcapinfo.c
@@ -307,7 +307,7 @@ main(int argc, char *argv[])
             }
 
             /* read the frame */
-            maxread = min(caplen, sizeof(buf));
+            maxread = min(caplen, (int)sizeof(buf));
             if ((ret = read(fd, &buf, maxread)) != maxread) {
                 if (ret < 0) {
                     printf("Error reading file: %s: %s\n", argv[i], strerror(errno));
